refactor: move magic number func to proper file
