+++ src/files/magic.rs
@@ -1,3 +1,5 @@
+use std::{fs::File, io::Read, path::PathBuf};
+
 // Magic number constants for file format detection
 
 #[cfg(target_os = "macos")]
@@ -26,3 +28,57 @@ pub const BZIP2_MAGIC: &[u8] = &[0x42, 0x5A, 0x68]; // "BZh"
 pub const TAR_MAGIC_OFFSET: usize = 257;
 pub const TAR_MAGIC: &[u8] = b"ustar";
 pub const SEVENZ_MAGIC: &[u8] = &[0x37, 0x7A, 0xBC, 0xAF, 0x27, 0x1C]; // 7z signature
+
+#[cfg(target_os = "linux")]
+fn is_exec_magic(buffer: &[u8; 4]) -> bool {
+    // Linux expects ELF binaries
+    buffer == &ELF_MAGIC // ELF
+}
+
+#[cfg(target_os = "windows")]
+fn is_exec_magic(buffer: &[u8; 4]) -> bool {
+    // Windows expects PE binaries (MZ header).
+    // Checking only the first two bytes because the other two may change,
+    // as they depend on the DOS stub.
+    buffer[..2] == core::magic::PE_MAGIC
+}
+
+#[cfg(target_os = "macos")]
+fn is_exec_magic(buffer: &[u8; 4]) -> bool {
+    // macOS expects Mach-O formats
+    MACHO_MAGIC_NUMBERS.contains(buffer)
+}
+
+#[cfg(not(target_os = "windows"))]
+pub fn is_exec_by_magic_number(path: &PathBuf) -> bool {
+    if let Ok(mut file) = File::open(path) {
+        let mut buffer = [0u8; 4];
+        if file.read_exact(&mut buffer).is_ok() {
+            return is_exec_magic(&buffer);
+        }
+    }
+    false
+}
+
+#[cfg(target_os = "windows")]
+pub fn is_exec_by_magic_number(path: &PathBuf) -> bool {
+    // We need to first check the file extension for Windows binaries,
+    // as it uses the PE format (MZ header) for file types other than
+    // .exe (e.g. .dll, .sys, etc.).
+    // Then we check the first two bytes of the .exe file because the
+    // other two may change (they depend on the DOS stub).
+    let extension = path
+        .extension()
+        .and_then(|s| s.to_str())
+        .unwrap_or_default();
+    if extension != "exe" {
+        return false;
+    }
+    if let Ok(mut file) = File::open(path) {
+        let mut buffer = [0u8; 4];
+        if file.read_exact(&mut buffer).is_ok() {
+            return is_exec_magic(&buffer);
+        }
+    }
+    false
+}
